+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Change Log
 
-All notable changes to the "ucx" extension will be documented in this file.
+## [0.11.0]
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+- added new option to override editor indentation style based on linter config
 
-## [Unreleased]
+## [0.10.2]
 
-- Initial release
+- better multiline variable declaration and multiline expressions
+- better operator spacing (fixed varios edge cases)
+
+## [0.10.1]
+
+- adds support for linter feature configuration support inside vscode
+
+## [0.9.7]
+
+- parser supports full 451b, 469b sources (no crash or error)